make tests nicer
